Commit 2
Run the TTVBotInitializer file to start up the bot then run TTVLOG.jar whenever you wish to view somones logs in your stream. 
In ttvbot.js username and password have to be changed to that of the account you want to use as the bot. Chanel & the var channelname need to be changed to that of the channel you want the bot to run on. Node.js & the extension tmi.js need to be downloaded (https://www.npmjs.com/package/tmi.js). Comes equipt with commands !social & !sr return values of these commands can be changed with a mod typing "!editcom com value". List of com's are fb, yt, tw, ig, & sr.